Move core methods (repair and normalRepair) from TestHelp to Main
