+++ src/test/java/ArrayHashCodeAndToStringProcessorTest.java
@@ -12,7 +12,7 @@ public class ArrayHashCodeAndToStringProcessorTest {
         String pathToRepairedFile = "./spooned/" + fileName;
 
         JavaCheckVerifier.verify(pathToBuggyFile, new ArrayHashCodeAndToStringCheck());
-        TestHelp.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,2116);
+        Main.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,2116);
         TestHelp.removeComplianceComments(pathToRepairedFile);
         JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new ArrayHashCodeAndToStringCheck());
     }
@@ -25,7 +25,7 @@ public class ArrayHashCodeAndToStringProcessorTest {
         String pathToRepairedFile = "./spooned/spoon/reflect/factory/" + fileName;
 
         JavaCheckVerifier.verify(pathToBuggyFile, new ArrayHashCodeAndToStringCheck());
-        TestHelp.repair(pathToBuggyFile,Constants.PROJECT_KEY,2116);
+        Main.repair(pathToBuggyFile,Constants.PROJECT_KEY,2116);
         TestHelp.removeComplianceComments(pathToRepairedFile);
         JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new ArrayHashCodeAndToStringCheck());
     }
