+++ src/test/java/BigDecimalDoubleConstructorProcessorTest.java
@@ -12,7 +12,7 @@ public class BigDecimalDoubleConstructorProcessorTest {
         String pathToRepairedFile = "./spooned/" + fileName;
 
         JavaCheckVerifier.verify(pathToBuggyFile, new BigDecimalDoubleConstructorCheck());
-        TestHelp.repair(pathToBuggyFile,Constants.PROJECT_KEY,2111);
+        Main.repair(pathToBuggyFile,Constants.PROJECT_KEY,2111);
         TestHelp.removeComplianceComments(pathToRepairedFile);
         JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new BigDecimalDoubleConstructorCheck());
     }
