+++ src/test/java/CompareStringsBoxedTypesWithEqualsProcessorTest.java
@@ -12,7 +12,7 @@ public class CompareStringsBoxedTypesWithEqualsProcessorTest {
         String pathToRepairedFile = "./spooned/" + fileName;
 
         JavaCheckVerifier.verify(pathToBuggyFile, new CompareStringsBoxedTypesWithEqualsCheck());
-        TestHelp.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,4973);
+        Main.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,4973);
         TestHelp.removeComplianceComments(pathToRepairedFile);
         JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new CompareStringsBoxedTypesWithEqualsCheck());
     }
