+++ src/test/java/UnclosedResourcesProcessorTest.java
@@ -12,7 +12,7 @@ public class UnclosedResourcesProcessorTest {
         String pathToRepairedFile = "./spooned/" + fileName;
 
         JavaCheckVerifier.verify(pathToBuggyFile, new UnclosedResourcesCheck());
-        TestHelp.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,2095);
+        Main.normalRepair(pathToBuggyFile,Constants.PROJECT_KEY,2095);
         TestHelp.removeComplianceComments(pathToRepairedFile);
         JavaCheckVerifier.verifyNoIssue(pathToRepairedFile, new UnclosedResourcesCheck());
